# api/openapi/v1/keyip.yaml
openapi: "3.0.3"
info:
  title: KeyIP-Intelligence API
  version: "1.0.0"
  description: |
    KeyIP-Intelligence is an AI-driven intellectual property lifecycle management platform
    designed for the OLED organic materials industry. It integrates molecular-level chemical
    understanding with patent law intelligence to support patent mining, infringement
    monitoring, portfolio optimization, and lifecycle management.

    ## Authentication

    All API endpoints require Bearer token authentication unless otherwise specified.
    Tokens are issued via the configured identity provider (Keycloak / OIDC-compatible).

    ## Pagination

    List endpoints support page-based pagination via `page` and `page_size` query parameters.
    Responses include pagination metadata.

    ## Error Handling

    All errors follow a unified structure:
    ```json
    {
      "code": "NOT_FOUND",
      "message": "The specified resource does not exist"
    }
    ```
  contact:
    name: KeyIP-Intelligence Open Source Community
    url: https://github.com/turtacn/KeyIP-Intelligence
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: https://api.keyip.example.com/api/v1
    description: Production server
  - url: https://staging-api.keyip.example.com/api/v1
    description: Staging server
  - url: http://localhost:8080
    description: Local development server
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/reports/fto:
    post:
      tags: [Reporting]
      summary: Generate FTO report
      description: Triggers asynchronous FTO (Freedom-to-Operate) report generation.
      operationId: generateFTOReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateFTOReportRequest"
      responses:
        "202":
          description: Report generation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/reports/infringement:
    post:
      tags: [Reporting]
      summary: Generate infringement report
      description: Triggers asynchronous infringement report generation.
      operationId: generateInfringementReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerateInfringementReportRequest"
      responses:
        "202":
          description: Report generation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/reports/portfolio:
    post:
      tags: [Reporting]
      summary: Generate portfolio report
      description: Triggers asynchronous portfolio report generation.
      operationId: generatePortfolioReport
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratePortfolioReportRequest"
      responses:
        "202":
          description: Report generation accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  report_id:
                    type: string
                  status:
                    type: string
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/reports:
    get:
      tags: [Reporting]
      summary: List reports
      operationId: listReports
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: status
          in: query
          schema:
            type: string
        - name: created_from
          in: query
          schema:
            type: string
            format: date-time
        - name: created_to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Report list
          content:
            application/json:
              schema:
                type: object
  /api/v1/reports/{report_id}/status:
    get:
      tags: [Reporting]
      summary: Get report generation status
      operationId: getReportStatus
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReportStatusResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports/{report_id}/download:
    get:
      tags: [Reporting]
      summary: Download report
      operationId: downloadReport
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
            enum: [pdf, docx]
            default: pdf
      responses:
        "200":
          description: Report file
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.wordprocessingml.document:
              schema:
                type: string
                format: binary
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Report not yet ready
  /api/v1/reports/{report_id}:
    delete:
      tags: [Reporting]
      summary: Delete report
      operationId: deleteReport
      parameters:
        - name: report_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Report deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/reports/templates:
    get:
      tags: [Reporting]
      summary: List report templates
      operationId: listReportTemplates
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: type
          in: query
          schema:
            type: string
        - name: format
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Template list
          content:
            application/json:
              schema:
                type: object
  /api/v1/reports/templates/{id}:
    get:
      tags: [Reporting]
      summary: Get report template
      operationId: getReportTemplate
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Template detail
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"

# =============================================================================
# Components
# =============================================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      description: Page number (starts at 1).
      schema:
        type: integer
        default: 1
        minimum: 1
    PageSize:
      name: page_size
      in: query
      description: Number of items per page (default 20, max 100).
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
...
  responses:
    BadRequest:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error description.

    # -------------------------------------------------------------------------
    # Health
    # -------------------------------------------------------------------------
...
    GenerateFTOReportRequest:
      type: object
      required: [target_smiles, jurisdiction]
      properties:
        target_smiles:
          type: string
        jurisdiction:
          type: string
        include_expired:
          type: boolean
          default: false
        depth:
          type: integer
          default: 2
        format:
          type: string
          enum: [pdf, docx, xlsx]
          default: pdf
        languages:
          type: array
          items:
            type: string
    GenerateInfringementReportRequest:
      type: object
      required: [patent_number, target_smiles]
      properties:
        patent_number:
          type: string
        target_smiles:
          type: array
          items:
            type: string
        analysis_depth:
          type: string
          enum: [literal, equivalents, comprehensive]
        format:
          type: string
          enum: [pdf, docx, xlsx]
          default: pdf
    GeneratePortfolioReportRequest:
      type: object
      required: [portfolio_id, report_type]
      properties:
        portfolio_id:
          type: string
        report_type:
          type: string
        format:
          type: string
          enum: [pdf, docx, xlsx]
          default: pdf
        include_charts:
          type: boolean
          default: false
    ReportStatusResponse:
      type: object
      properties:
        report_id:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        progress:
          type: number
          format: double
        report_type:
          type: string
        format:
          type: string
        created_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        error:
          type: string
        download_url:
          type: string
          format: uri